+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
 - Entities Paginations (#113)
 - Table pagination && elements configurable by env
 - Resource type pagination (#109)
+- Policies Paginations (#127)
 
 ### Bug fixes